cfi_flash: fix writing data to flash from an unaligned source
The cfi_flash driver expects source data being written to flash are properly
aligned. In some cases, such as TFTP/NFS transfer directly to flash, the data
are passed to the driver from the network packet buffer and can be unaligned.
Access to unaligned data crashes U-Boot on architectures, which don't support
unaligned data access.

To work around this issue new setting CONFIG_SYS_FLASH_WRITE_UNALIGNED
is implemented. This setting enables accurate access to source data using the
get_unaligned macro.

Note that there are no noticeable performance penalties as the result,
so as writing to flash is the significantly slower operation.

Signed-off-by: Igor A. Perminov <igor.a.perminov@gmail.com>