tion270: fixed USB MFP setup
USB MFPs (88, 89) were configured unconditionally via GAFR at the board
startup time and were at this state, when Linux started. This state
caused malfunction of the USB host port 1 in Linux 2.6.35, so as
the pin 88 isn't reconfigured as the GPIO there.
Fixed: MFPs 88 & 89 are configured for USB in the USB startup routine
and deconfigured back to GPIO in the USB stop routine.

Signed-off-by: Igor A. Perminov <igor.a.perminov@gmail.com>